Move forkawesome, colorspacehandling, fix .xyz extension method
