+++ src/StudioCore/ParamEditor/AutoFill.cs
@@ -1,6 +1,7 @@
 ﻿using Andre.Formats;
 using static Andre.Native.ImGuiBindings;
 using StudioCore.Editor;
+using StudioCore.Utilities;
 using System;
 using System.Collections.Generic;
 using System.Linq;
